Added 2 more views
View for Tables works flawless. You can change from table view to OrderView and back without any problems.
The OrderView can remove orders with the possibility to recover the order. 
Loginview are created but nothing in it yet. 

The android manifest file are setup to work with Order and table view.
Changed color of the buttons